Minor fixes for tests: 1. Verification of array content 2. Avoid using Array.Empty which was added in later version of contracts 3. Moving tests involving types that are not in public contracts to XmlSerializerTests.Internal.cs
